backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
dnspython==2.6.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.124.4
fastapi-cli==0.0.16
fastapi-cloud-cli==0.7.0
fastar==0.8.0
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
google-ai-generativelanguage==0.1.0
google-api-core==2.28.1
google-auth==2.45.0
google-generativeai==0.1.0rc1
googleapis-common-protos==1.72.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.11
importlib_resources==6.4.5
Jinja2==3.1.6
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
packaging==25.0
pillow==10.4.0
platformdirs==4.3.6
polars==1.8.2
proto-plus==1.27.0
protobuf==4.25.8
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.4
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rsa==4.9.1
scenedetect==0.6.7.1
scipy==1.10.1
sentry-sdk==2.48.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.44.0
sympy==1.13.3
tomli==2.3.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
triton==3.0.0
typer==0.20.1
typing_extensions==4.13.2
ultralytics==8.3.240
ultralytics-thop==2.0.18
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.22.1
watchfiles==0.24.0
websockets==13.1
zipp==3.20.2
